ld: support -dynamic_lookup_library
This causes all symbols found in the specified library to be looked up
at load time. The library itself is not linked by the output library at
all.

The main use case for this is creating compiled modules for dynamic
languages (e.g., Ruby or Python) where the application loading the
module provides the interpreter's symbols. This allows the compiled
module to be agnostic to the installation of the interpreter. That is,
it can link against the Apple-provided Python, but be run with an
Anaconda-provided interpreter.